+++ mutable/set.go
@@ -5,20 +5,20 @@ import (
 	"sort"
 
 	"github.com/m4gshm/gollections/c"
-	"github.com/m4gshm/gollections/it/impl/it"
+	"github.com/m4gshm/gollections/iter/impl/iter"
+	"github.com/m4gshm/gollections/loop"
 	"github.com/m4gshm/gollections/map_"
 	"github.com/m4gshm/gollections/mutable/ordered"
-	"github.com/m4gshm/gollections/predicate"
 	"github.com/m4gshm/gollections/slice"
 )
 
-// NewSet creates a set with a predefined capacity.
-func NewSet[T comparable](capacity int) Set[T] {
+// NewSetCap creates a set with a predefined capacity.
+func NewSetCap[T comparable](capacity int) *Set[T] {
 	return WrapSet(make(map[T]struct{}, capacity))
 }
 
-// ToSet converts an elements slice to the set containing them.
-func ToSet[T comparable](elements []T) Set[T] {
+// NewSet instantiates Set and copies elements to it.
+func NewSet[T comparable](elements []T) *Set[T] {
 	internal := make(map[T]struct{}, len(elements))
 	for _, v := range elements {
 		internal[v] = struct{}{}
@@ -26,83 +26,138 @@ func ToSet[T comparable](elements []T) Set[T] {
 	return WrapSet(internal)
 }
 
+// ToSet creates a Set instance with elements obtained by passing an iterator.
+func ToSet[T comparable](elements c.Iterator[T]) *Set[T] {
+	internal := map[T]struct{}{}
+	if elements != nil {
+		for e, ok := elements.Next(); ok; e, ok = elements.Next() {
+			internal[e] = struct{}{}
+		}
+	}
+	return WrapSet(internal)
+}
+
 // WrapSet creates a set using a map as the internal storage.
-func WrapSet[K comparable](elements map[K]struct{}) Set[K] {
-	return Set[K]{elements: elements}
+func WrapSet[T comparable](elements map[T]struct{}) *Set[T] {
+	return &Set[T]{elements: elements}
 }
 
 // Set is the Collection implementation that provides element uniqueness. The elements must be comparable.
-type Set[K comparable] struct {
-	elements map[K]struct{}
+type Set[T comparable] struct {
+	elements map[T]struct{}
 }
 
 var (
 	_ c.Addable[int]          = (*Set[int])(nil)
 	_ c.AddableNew[int]       = (*Set[int])(nil)
+	_ c.AddableAll[int]       = (*Set[int])(nil)
+	_ c.AddableAllNew[int]    = (*Set[int])(nil)
 	_ c.Deleteable[int]       = (*Set[int])(nil)
 	_ c.DeleteableVerify[int] = (*Set[int])(nil)
 	_ c.Set[int]              = (*Set[int])(nil)
 	_ fmt.Stringer            = (*Set[int])(nil)
-	_ c.Addable[int]          = Set[int]{}
-	_ c.AddableNew[int]       = Set[int]{}
-	_ c.Deleteable[int]       = Set[int]{}
-	_ c.DeleteableVerify[int] = Set[int]{}
-	_ c.Set[int]              = Set[int]{}
-	_ fmt.Stringer            = Set[int]{}
+	// _ c.Addable[int]          = Set[int]{}
+	// _ c.AddableNew[int]       = Set[int]{}
+	// _ c.AddableAll[int]       = Set[int]{}
+	// _ c.AddableAllNew[int]    = Set[int]{}
+	// _ c.Deleteable[int]       = Set[int]{}
+	// _ c.DeleteableVerify[int] = Set[int]{}
+	// _ c.Set[int]              = Set[int]{}
+	// _ fmt.Stringer            = Set[int]{}
 )
 
-func (s Set[K]) Begin() c.Iterator[K] {
-	return s.Head()
+func (s *Set[T]) Begin() c.Iterator[T] {
+	h := s.Head()
+	return &h
 }
 
-func (s Set[K]) BeginEdit() c.DelIterator[K] {
-	return s.Head()
+func (s *Set[T]) BeginEdit() c.DelIterator[T] {
+	h := s.Head()
+	return &h
 }
 
-func (s Set[K]) Head() *SetIter[K] {
+func (s *Set[T]) Head() SetIter[T] {
+	if s == nil || s.elements == nil {
+		return NewSetIter[T](nil, nil)
+	}
 	return NewSetIter(s.elements, s.DeleteOne)
 }
 
-func (s Set[K]) Collect() []K {
+func (s *Set[T]) First() (SetIter[T], T, bool) {
+	var (
+		iterator  = s.Head()
+		first, ok = iterator.Next()
+	)
+	return iterator, first, ok
+}
+
+func (s *Set[T]) Slice() []T {
+	if s == nil || s.elements == nil {
+		return nil
+	}
 	return map_.Keys(s.elements)
 }
 
-func (s Set[T]) Copy() Set[T] {
+func (s *Set[T]) Copy() *Set[T] {
+	if s == nil || s.elements == nil {
+		return nil
+	}
 	return WrapSet(map_.Clone(s.elements))
 }
 
-func (s Set[T]) IsEmpty() bool {
+func (s *Set[T]) IsEmpty() bool {
 	return s.Len() == 0
 }
 
-func (s Set[K]) Len() int {
+func (s *Set[T]) Len() int {
+	if s == nil || s.elements == nil {
+		return 0
+	}
 	return len(s.elements)
 }
 
-func (s Set[K]) Contains(val K) bool {
+func (s *Set[T]) Contains(val T) bool {
+	if s == nil || s.elements == nil {
+		return false
+	}
 	_, ok := s.elements[val]
 	return ok
 }
 
-func (s Set[K]) Add(elements ...K) {
+func (s *Set[T]) Add(elements ...T) {
+	if s == nil {
+		return
+	}
 	for _, element := range elements {
 		s.AddOne(element)
 	}
 }
 
-func (s Set[K]) AddOne(element K) {
-	s.elements[element] = struct{}{}
+func (s *Set[T]) AddOne(element T) {
+	if s == nil {
+		return
+	} else if elements := s.elements; elements == nil {
+		s.elements = map[T]struct{}{element: {}}
+	} else {
+		elements[element] = struct{}{}
+	}
 }
 
-func (s Set[K]) AddNew(elements ...K) bool {
+func (s *Set[T]) AddNew(elements ...T) bool {
+	if s == nil {
+		return false
+	}
 	ok := false
 	for _, element := range elements {
-		ok = s.AddNewOne(element) || ok
+		ok = s.AddOneNew(element) || ok
 	}
 	return ok
 }
 
-func (s Set[K]) AddNewOne(element K) bool {
+func (s *Set[T]) AddOneNew(element T) bool {
+	if s == nil {
+		return false
+	}
 	ok := !s.Contains(element)
 	if ok {
 		s.elements[element] = struct{}{}
@@ -110,17 +165,42 @@ func (s Set[K]) AddNewOne(element K) bool {
 	return ok
 }
 
-func (s Set[K]) Delete(elements ...K) {
+func (s *Set[T]) AddAll(elements c.Iterable[T]) {
+	if s == nil {
+		return
+	}
+	loop.ForEach(elements.Begin().Next, s.AddOne)
+}
+
+func (s *Set[T]) AddAllNew(elements c.Iterable[T]) bool {
+	if s == nil {
+		return false
+	}
+	var ok bool
+	loop.ForEach(elements.Begin().Next, func(v T) { ok = s.AddOneNew(v) || ok })
+	return ok
+}
+
+func (s *Set[T]) Delete(elements ...T) {
+	if s == nil {
+		return
+	}
 	for _, element := range elements {
 		s.DeleteOne(element)
 	}
 }
 
-func (s Set[K]) DeleteOne(element K) {
+func (s *Set[T]) DeleteOne(element T) {
+	if s == nil {
+		return
+	}
 	delete(s.elements, element)
 }
 
-func (s Set[T]) DeleteActual(elements ...T) bool {
+func (s *Set[T]) DeleteActual(elements ...T) bool {
+	if s == nil {
+		return false
+	}
 	ok := false
 	for i := range elements {
 		ok = s.DeleteActualOne(elements[i]) || ok
@@ -128,7 +208,10 @@ func (s Set[T]) DeleteActual(elements ...T) bool {
 	return ok
 }
 
-func (s Set[K]) DeleteActualOne(element K) bool {
+func (s *Set[T]) DeleteActualOne(element T) bool {
+	if s == nil || s.elements == nil {
+		return false
+	}
 	_, ok := s.elements[element]
 	if ok {
 		delete(s.elements, element)
@@ -136,24 +219,32 @@ func (s Set[K]) DeleteActualOne(element K) bool {
 	return ok
 }
 
-func (s Set[K]) For(walker func(K) error) error {
+func (s *Set[T]) For(walker func(T) error) error {
+	if s == nil {
+		return nil
+	}
 	return map_.ForKeys(s.elements, walker)
 }
 
-func (s Set[K]) ForEach(walker func(K)) {
-	map_.ForEachKey(s.elements, walker)
+func (s *Set[T]) ForEach(walker func(T)) {
+	if s != nil {
+		map_.ForEachKey(s.elements, walker)
+	}
 }
 
-func (s Set[K]) Filter(filter predicate.Predicate[K]) c.Pipe[K, []K] {
-	return it.NewPipe[K](it.Filter(s.Head(), filter))
+func (s *Set[T]) Filter(filter func(T) bool) c.Pipe[T] {
+	h := s.Head()
+	return iter.NewPipe[T](iter.Filter(h, h.Next, filter))
 }
 
-func (s Set[K]) Map(by c.Converter[K, K]) c.Pipe[K, []K] {
-	return it.NewPipe[K](it.Map(s.Head(), by))
+func (s *Set[T]) Convert(by func(T) T) c.Pipe[T] {
+	h := s.Head()
+	return iter.NewPipe[T](iter.Convert(h, h.Next, by))
 }
 
-func (s Set[K]) Reduce(by c.Binary[K]) K {
-	return it.Reduce(s.Head(), by)
+func (s *Set[T]) Reduce(by func(T, T) T) T {
+	h := s.Head()
+	return loop.Reduce(h.Next, by)
 }
 
 // Sort transforms to the ordered Set contains sorted elements.
@@ -166,11 +257,11 @@ func (s *Set[T]) StableSort(less slice.Less[T]) *ordered.Set[T] {
 }
 
 func (s *Set[T]) sortBy(sorter slice.Sorter, less slice.Less[T]) *ordered.Set[T] {
-	c := slice.Clone(s.Collect())
-	slice.Sort(c, sorter, less)
-	return ordered.ToSet(c)
+	cl := slice.Clone(s.Slice())
+	slice.Sort(cl, sorter, less)
+	return ordered.NewSet(cl)
 }
 
-func (s Set[K]) String() string {
-	return slice.ToString(s.Collect())
+func (s *Set[T]) String() string {
+	return slice.ToString(s.Slice())
 }
